Fix debug configuration querying not inheriting environment
